Persist data during "Given a conference named" task"
In order to replace the static implementation with the real one the
"Given" task needs to persist actual data. We try use domain language
event for this operation: in the real world a Conference is not
"persisted", it is "Planned", so we update the step according to reflect
this language and we implement a method planConference() which has the
responsibility to store data into the database.